nothin

+++ nicegrill/modules/google.py
@@ -1,5 +1,6 @@
 import logging
 import os
+from PIL import Image
 from nicegrill import utils
 from search_engine_parser import GoogleSearch
 from google_images_download import google_images_download
@@ -20,7 +21,7 @@ class Google:
             pass
         await message.edit(text, link_preview=False)
 
-    async def gphotoxxx(message):
+    async def gimgxxx(message):
         keyword = utils.get_arg(message)
         response = google_images_download.googleimagesdownload()
         args = {"keywords": keyword, "limit":5, "print_urls":False}
@@ -30,7 +31,17 @@ class Google:
             await message.edit("<i>Nothing found</i>")
             return
         await message.edit("<i>Uploading..</i>")
-        await message.client.send_message(message.chat_id, file=paths[0][keyword])
+        newlist = []
+        for filename in os.listdir(f"downloads/{keyword}/"):
+            try:
+                with Image.open(f"downloads/{keyword}/{filename}") as im:
+                    newlist.append(f"downloads/{keyword}/{filename}")
+            except Exception:
+                os.remove(f"downloads/{keyword}/{filename}")
+        if not newlist:
+            await message.edit("<i>Images were broken so nothing happened</i>")
+            return
+        await message.client.send_message(message.chat_id, file=newlist)
         await message.delete()
         for path in paths[0][keyword]:
             os.remove(path)
